migrate t5 to hf
